fix: Move `fetch<db_table>` to client-side
