+++ src/actions/attendance.ts
@@ -23,20 +23,6 @@ export async function registerPlayer(
     throw new Error(`registerPlayer encountered an error: ${error.message}`);
 }
 
-export async function fetchAttendance(session: Session): Promise<Attendance[]> {
-  const supabase = await createClient();
-
-  const { data: attendance, error } = await supabase
-    .from("attendance")
-    .select("*")
-    .eq("session_id", session.id);
-
-  if (error)
-    throw new Error(`fetchAttendance encountered an error: ${error.message}`);
-
-  return attendance ?? [];
-}
-
 export async function updateAttendance(
   session: Session,
   player: Player,
