+++ src/actions/logs.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { createClient } from "@/lib/supabase/server";
-import type { Log, Tournament } from "@/lib/types";
+import type { Log } from "@/lib/types";
 
 export async function createLog(log: Log): Promise<void> {
   const supabase = await createClient();
@@ -10,20 +10,6 @@ export async function createLog(log: Log): Promise<void> {
 
   if (error)
     throw new Error(`createLog encountered an error: ${error.message}`);
-}
-
-export async function fetchLogs(tournament: Tournament): Promise<Log[]> {
-  const supabase = await createClient();
-
-  const { data: logs, error } = await supabase
-    .from("logs")
-    .select("*")
-    .eq("tournament_id", tournament.id);
-
-  if (error)
-    throw new Error(`fetchLogs encountered an error: ${error.message}`);
-
-  return logs ?? [];
 }
 
 export async function disableLog(log: Log): Promise<void> {
