+++ src/actions/players.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { createClient } from "@/lib/supabase/server";
-import type { Player, Tournament } from "@/lib/types";
+import type { Player } from "@/lib/types";
 
 export async function createPlayer(player: Player): Promise<void> {
   const supabase = await createClient();
@@ -10,20 +10,6 @@ export async function createPlayer(player: Player): Promise<void> {
 
   if (error)
     throw new Error(`createPlayer encountered an error: ${error.message}`);
-}
-
-export async function fetchPlayers(tournament: Tournament): Promise<Player[]> {
-  const supabase = await createClient();
-
-  const { data: players, error } = await supabase
-    .from("players")
-    .select("*")
-    .eq("tournament_id", tournament.id);
-
-  if (error)
-    throw new Error(`fetchPlayers encountered an error: ${error.message}`);
-
-  return players ?? [];
 }
 
 export async function updatePlayer(player: Player): Promise<void> {
