+++ src/actions/sessions.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { createClient } from "@/lib/supabase/server";
-import type { Session, Tournament } from "@/lib/types";
+import type { Session } from "@/lib/types";
 
 export async function createSession(session: Session): Promise<void> {
   const supabase = await createClient();
@@ -10,20 +10,4 @@ export async function createSession(session: Session): Promise<void> {
 
   if (error)
     throw new Error(`createSession encountered an error: ${error.message}`);
-}
-
-export async function fetchSessions(
-  tournament: Tournament,
-): Promise<Session[]> {
-  const supabase = await createClient();
-
-  const { data: sessions, error } = await supabase
-    .from("sessions")
-    .select("*")
-    .eq("tournament_id", tournament.id);
-
-  if (error)
-    throw new Error(`fetchSessions encountered an error: ${error.message}`);
-
-  return sessions ?? [];
 }
