+++ src/actions/tournaments.ts
@@ -1,9 +1,8 @@
 "use server";
 
 import { createSession } from "@/actions/sessions";
-import type { Tables } from "@/lib/database.types";
 import { createClient } from "@/lib/supabase/server";
-import type { ScoringRule, Tournament } from "@/lib/types";
+import type { ScoringRule } from "@/lib/types";
 
 export async function createTournament(
   tournamentName: string,
@@ -28,26 +27,6 @@ export async function createTournament(
   };
 
   await createSession(initialSession);
-}
-
-// Supabase stores `scoring_rules` as jsonb so we need to override it with the correct type
-type TournamentRow = Omit<Tables<"tournaments">, "scoring_rules"> & {
-  scoring_rules: ScoringRule[];
-};
-
-export async function fetchTournaments(): Promise<Tournament[]> {
-  const supabase = await createClient();
-
-  const { data: tournaments, error } = await supabase
-    .from("tournaments")
-    .select("*")
-    .order("last_updated", { ascending: false })
-    .overrideTypes<TournamentRow[]>(); // Not ideal, but otherwise the typing breaks :(
-
-  if (error)
-    throw new Error(`fetchTournaments encountered an error: ${error.message}`);
-
-  return tournaments ?? [];
 }
 
 export async function getTournamentName(tournamentId: string): Promise<string> {
