+++ src/hooks/tournaments/useTournaments.ts
@@ -1,6 +1,7 @@
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { useCallback } from "react";
-import { fetchTournaments } from "@/actions/tournaments";
+import type { Tables } from "@/lib/database.types";
+import { createClient } from "@/lib/supabase/client";
 import type { ScoringRule, Tournament } from "@/lib/types";
 import { useTournamentContext } from "@/providers/TournamentProvider";
 
@@ -37,4 +38,24 @@ export function useTournaments(): UseTournamentsType {
     currentTournament: query.data.currentTournament,
     scoringRules: query.data.scoringRules,
   };
+}
+
+// Supabase stores `scoring_rules` as jsonb so we need to override it with the correct type
+type TournamentRow = Omit<Tables<"tournaments">, "scoring_rules"> & {
+  scoring_rules: ScoringRule[];
+};
+
+export async function fetchTournaments(): Promise<Tournament[]> {
+  const supabase = createClient();
+
+  const { data: tournaments, error } = await supabase
+    .from("tournaments")
+    .select("*")
+    .order("last_updated", { ascending: false })
+    .overrideTypes<TournamentRow[]>(); // Not ideal, but otherwise the typing breaks :(
+
+  if (error)
+    throw new Error(`fetchTournaments encountered an error: ${error.message}`);
+
+  return tournaments ?? [];
 }
